Move tests for pylint_plugins
