+++ pylint_plugins/tests/test_pytest_raises_checker.py
@@ -1,17 +1,14 @@
 import pytest
+import astroid
+from pylint.testutils import MessageTest
+from pylint.testutils import CheckerTestCase
 
-from tests.pylint_plugins.utils import create_message, extract_node, skip_if_pylint_unavailable
-
-pytestmark = skip_if_pylint_unavailable()
+from pylint_plugins import PytestRaisesChecker
 
 
 @pytest.fixture(scope="module")
 def test_case():
-    # Ref: https://pylint.pycqa.org/en/latest/how_tos/custom_checkers.html#testing-a-checker
-    import pylint.testutils
-    from pylint_plugins import PytestRaisesChecker
-
-    class TestPytestRaisesWithoutMatch(pylint.testutils.CheckerTestCase):
+    class TestPytestRaisesWithoutMatch(CheckerTestCase):
         CHECKER_CLASS = PytestRaisesChecker
 
     test_case = TestPytestRaisesWithoutMatch()
@@ -21,7 +18,7 @@ def test_case():
 
 def without_raises_bad_cases():
     # Single context manager
-    root_node = extract_node(
+    root_node = astroid.extract_node(
         """
 with pytest.raises(Exception):
     raise Exception("failed")
@@ -30,7 +27,7 @@ with pytest.raises(Exception):
     yield root_node, root_node.items[0][0], (2, 5)
 
     # Multiple context managers
-    root_node = extract_node(
+    root_node = astroid.extract_node(
         """
 with context_manager, pytest.raises(Exception):
     raise Exception("failed")
@@ -39,7 +36,7 @@ with context_manager, pytest.raises(Exception):
     yield root_node, root_node.items[1][0], (2, 22)
 
     # Without `with`
-    root_node = extract_node(
+    root_node = astroid.extract_node(
         """
 pytest.raises(Exception)
 """
@@ -49,7 +46,7 @@ pytest.raises(Exception)
 
 def without_raises_iter_good_cases():
     # Single context manager
-    yield extract_node(
+    yield astroid.extract_node(
         """
 with pytest.raises(Exception, match="failed"):
     raise Exception("failed")
@@ -57,7 +54,7 @@ with pytest.raises(Exception, match="failed"):
     )
 
     # Multiple context managers
-    yield extract_node(
+    yield astroid.extract_node(
         """
 with context_manager, pytest.raises(Exception, match="failed"):
     raise Exception("failed")
@@ -65,7 +62,7 @@ with context_manager, pytest.raises(Exception, match="failed"):
     )
 
     # Without `with`
-    yield extract_node(
+    yield astroid.extract_node(
         """
 pytest.raises(Exception, match="failed")
 """
@@ -75,7 +72,12 @@ pytest.raises(Exception, match="failed")
 def test_without_raises(test_case):
     for root_node, error_node, (line, col_offset) in without_raises_bad_cases():
         with test_case.assertAddsMessages(
-            create_message(test_case.CHECKER_CLASS.WITHOUT_MATCH, error_node, line, col_offset)
+            MessageTest(
+                test_case.CHECKER_CLASS.WITHOUT_MATCH,
+                node=error_node,
+                line=line,
+                col_offset=col_offset,
+            )
         ):
             test_case.walk(root_node)
 
@@ -85,7 +87,7 @@ def test_without_raises(test_case):
 
 
 def multiple_statements_bad_cases():
-    root_node = extract_node(
+    root_node = astroid.extract_node(
         """
 with pytest.raises(Exception, match="failed"):
     a = 1
@@ -95,7 +97,7 @@ with pytest.raises(Exception, match="failed"):
     )
     yield root_node, root_node, (2, 0)
 
-    root_node = extract_node(
+    root_node = astroid.extract_node(
         """
 with pytest.raises(Exception, match="failed"), mock_patch("module.function"):
     a = 1
@@ -107,14 +109,14 @@ with pytest.raises(Exception, match="failed"), mock_patch("module.function"):
 
 
 def multiple_statements_good_cases():
-    yield extract_node(
+    yield astroid.extract_node(
         """
 with pytest.raises(Exception, match="failed"):
     function_that_throws()
 """
     )
 
-    yield extract_node(
+    yield astroid.extract_node(
         """
 with pytest.raises(Exception, match="failed"):
     if condition:
@@ -128,8 +130,11 @@ with pytest.raises(Exception, match="failed"):
 def test_multiple_statements(test_case):
     for root_node, error_node, (line, col_offset) in multiple_statements_bad_cases():
         with test_case.assertAddsMessages(
-            create_message(
-                test_case.CHECKER_CLASS.MULTIPLE_STATEMENTS, error_node, line, col_offset
+            MessageTest(
+                test_case.CHECKER_CLASS.MULTIPLE_STATEMENTS,
+                node=error_node,
+                line=line,
+                col_offset=col_offset,
             )
         ):
             test_case.walk(root_node)
